+++ JDI.Light/JDI.Light.Tests/JDI.Light.Tests.csproj
@@ -66,7 +66,7 @@
     <Compile Include="Enums\Odds.cs" />
     <Compile Include="Enums\Weather.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Tests\Common\CheckListTests.cs" />
+    <Compile Include="Tests\Complex\CheckListTests.cs" />
     <Compile Include="Tests\Common\ComboBoxTests.cs" />
     <Compile Include="Tests\Common\DateTimeTests.cs" />
     <Compile Include="Tests\Common\FileInputTests.cs" />
