+++ JDI.Light/JDI.Light/Settings/MapInterfaceToElement.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using JDI.Light.Elements.Base;
 using JDI.Light.Elements.Common;
+using JDI.Light.Elements.Complex;
 using JDI.Light.Interfaces.Base;
 using JDI.Light.Interfaces.Common;
 using OpenQA.Selenium;
